+++ wpull/url.py
@@ -5,6 +5,7 @@ import collections
 import fnmatch
 import functools
 import itertools
+import namedlist
 import re
 import string
 import sys
@@ -30,10 +31,25 @@ _URLInfoType = collections.namedtuple(
         'hostname',
         'port',
         'raw',
-        'url',
         'encoding',
     ]
 )
+
+
+NormalizationParams = namedlist.namedtuple(
+    'NormalizationParamsType',
+    [
+        ('sort_query', False),
+        ('always_delim_query', False)
+    ]
+)
+'''Parameters for URL normalization.
+
+Args:
+    sort_query (bool): Whether to sort the query string items.
+    always_delim_query: Whether to always deliminate the key-value items where
+        value is empty.
+'''
 
 
 class URLInfo(_URLInfoType):
@@ -70,8 +86,17 @@ class URLInfo(_URLInfoType):
     }
 
     @classmethod
-    def parse(cls, string, default_scheme='http', encoding='utf8'):
+    def parse(cls, string, default_scheme='http', encoding='utf8',
+    normalization_params=None):
         '''Parse and return a new info from the given URL.
+
+        Args:
+            string (str): The URL.
+            default_scheme (str): The default scheme if not specified.
+            encoding (str): The name of the encoding to be used for IRI
+                support.
+            normalization_params: Instance of :class:`NormalizationParams`
+                describing further normalization settings.
 
         Returns:
             :class:`URLInfo`
@@ -83,6 +108,9 @@ class URLInfo(_URLInfoType):
             raise ValueError('Empty URL')
 
         assert isinstance(string, str)
+
+        if normalization_params is None:
+            normalization_params = NormalizationParams()
 
         url_split_result = urllib.parse.urlsplit(string)
 
@@ -109,43 +137,19 @@ class URLInfo(_URLInfoType):
             url_split_result.scheme,
             url_split_result.netloc,
             cls.normalize_path(url_split_result.path, encoding=encoding),
-            cls.normalize_query(url_split_result.query, encoding=encoding),
+            cls.normalize_query(
+                url_split_result.query, encoding=encoding,
+                sort=normalization_params.sort_query,
+                always_delim=normalization_params.always_delim_query,
+            ),
             url_split_result.fragment,
             url_split_result.username,
             url_split_result.password,
             cls.normalize_hostname(url_split_result.hostname),
             port,
             string,
-            cls.normalize(url_split_result, encoding=encoding),
             wpull.util.normalize_codec_name(encoding),
         )
-
-    @classmethod
-    def normalize(cls, url_split_result, encoding='utf8'):
-        '''Return a normalized URL string.'''
-        if url_split_result.scheme not in ('http', 'https'):
-            return url_split_result.geturl()
-
-        default_port = cls.DEFAULT_PORTS.get(url_split_result.scheme)
-
-        if default_port == url_split_result.port:
-            host_with_port = cls.normalize_hostname(url_split_result.hostname)
-        else:
-            host_with_port = url_split_result.netloc.split('@', 1)[-1]
-            if ':' in host_with_port:
-                host, port = host_with_port.rsplit(':', 1)
-                host_with_port = '{0}:{1}'.format(
-                    cls.normalize_hostname(host), port)
-            else:
-                host_with_port = cls.normalize_hostname(host_with_port)
-
-        return urllib.parse.urlunsplit([
-            url_split_result.scheme,
-            host_with_port,
-            cls.normalize_path(url_split_result.path, encoding=encoding),
-            cls.normalize_query(url_split_result.query, encoding=encoding),
-            ''
-        ])
 
     @classmethod
     def normalize_hostname(cls, hostname):
@@ -179,8 +183,8 @@ class URLInfo(_URLInfoType):
             query (str): The query string.
             encoding (str): IRI encoding.
             sort (bool): If True, the items will be sorted.
-            always_delim (bool): If True, the equal sign ``=`` will always be
-                present for each item.
+            always_delim (bool): If True, the equal sign ``=`` deliminator
+                will always be present for each key-value item.
         '''
         if not query:
             return
@@ -207,6 +211,21 @@ class URLInfo(_URLInfoType):
             if value is not None or always_delim else
             quote_func(name)
             for name, value in query_list])
+
+    @property
+    def url(self):
+        '''Return a normalized URL string.'''
+        if self.scheme not in ('http', 'https'):
+            url_split_result = urllib.parse.urlsplit(self.raw)
+            return url_split_result.geturl()
+
+        return urllib.parse.urlunsplit([
+            self.scheme,
+            self.hostname_with_port,
+            self.path,
+            self.query,
+            ''
+        ])
 
     def is_port_default(self):
         if self.scheme in self.DEFAULT_PORTS:
@@ -494,6 +513,19 @@ class BackwardFilenameFilter(BaseURLFilter):
             match = re.search(fnmatch.translate(suffix), test_filename)
             if match:
                 return True
+
+
+def normalize(url, **kwargs):
+    '''Normalize a URL.
+
+    This function is a convenience function that is equivalent to::
+
+        >>> URLInfo.parse('http://example.com').url
+        'http://example.com'
+
+    :seealso: :func:`URLInfo.parse`.
+    '''
+    return URLInfo.parse(url, **kwargs).url
 
 
 RELAXED_SAFE_CHARS = '/!$&()*+,:;=@~'